refactor: reorganize and remove unused Escuela components
- Removed multiple unused components related to the Escuela feature, including pages for `escuela`, `ciclosEscolares`, `materias`, `padres`, `salones`, and others.
- Updated import paths for `CrudDialog` and other components to reflect the new structure.
- Cleaned up the sidebar navigation to remove references to deleted pages and ensure proper links to existing features.
- This restructuring aims to simplify the codebase and improve maintainability.
