+++ apps/system/app/[subdomain]/(dashboard)/catalogoDeClases/page.tsx
