+++ apps/system/app/[subdomain]/(dashboard)/ciclosEscolares/[id]/page.tsx
