+++ apps/system/app/[subdomain]/(dashboard)/ciclosEscolares/page.tsx
