+++ apps/system/app/[subdomain]/(dashboard)/clasesPorAlumnos/assignacion-masiva/page.tsx
