+++ apps/system/app/[subdomain]/(dashboard)/departamentos/page.tsx
