+++ apps/system/app/[subdomain]/(dashboard)/eventosPorClase/page.tsx
