+++ apps/system/app/[subdomain]/(dashboard)/periodosPorClase/page.tsx
