+++ apps/system/app/[subdomain]/(dashboard)/public/nosotros/page.tsx
