+++ apps/system/app/[subdomain]/(dashboard)/salones/page.tsx
